expo: improve support for non-square workspace grids
Up to now, when vwidth != vheight, we simply used different vertical and
horizontal scaling. However, the result is quite ugly and unusable. Now
we use the same scaling which is taken by the max(vwith, vheight) and
the final image is shifted so that the grid appears in the middle. All
relevant functions - moving a view and clicking on a workspace work as
expected