animation: re-implement fire animation
The code was re-adapted from the initial implementation for X11.
Currently it requires compute shaders(GLES 3.1). In the future there
should be also a "software" version of the particle system it uses.

Other changes:
* Change the way desktop_surface_removed is handled. We don't remove the
view if any plugin wants to keep it(view->keep_count > 0). Instead, we
keep it in the renderable views list and set its destroyed flag to true.
All plugins and the core should check for this type of views and ignore
them.
* Introduce the wf_abilites enum in wayfire_grab_interface. This is
intended as an upgrade to the previous compat/compatAll fields. Using
wf_abilities a plugin can specify which functions it uses(for ex. custom
rendering). output->activate_plugin() cannot activate two plugins with
the same ability. This allows for example the screenshot plugin and
others to work at the same time, while expo and cube cannot due to the
same abilities mask.
* Implement fade animation as a animation_base. While using the weston
built-in function for fade animation is shorter, the custom
implementation fits into the animation_hook framework, which all other
animations (will) use