backlight: add min_brightess setting
Currently, we set the backlight to 0 when sleeping. However,
on some laptops with touchscreen it gets disabled so that the compositor
cannot wake unless a keyboard/pointer has been used. In this case, the
user can specify a minimal brightness to be set at such events. We must
also make sure to delay backlight restoration, as it might interfere
with the system-fade-animation