shell: implement sound-popup
It is implemented as a standalone binary. It supports touch/pointer
input to adjust audio, as well reads current status. If another instance
is started while the first one is running, the first one will refresh
and the second will terminate. It also accepts up to 5 parameters:
wayfire-sound-popup x y w h t

x, y, w, h describe the geometry of the popup to be created on the
current output. t is the amount of time spend idle before closing the
popup. With these options it can be used to control the volume from the
panel(as a launcher) and also as a frontend for a keybinging(see sample
        configuration for more details)