workspace-manager: implement fullscreen layer
This implements a fullscreen layer. Fullscreen views are supposed to be
rendered (when focused) in their own layer, which sits between the top
and the overlay layers. However, when they aren't the topmost in stack,
they should be put back into the workspace layer.

The current implementation tries to keep changes to a minimum. To ensure
the fullscreen layer is properly maintained, the current implementation
adjusts it whenever a view's layer changes. Here's what we need to do:

1) When a window is focused, if it is fullscreen, just add it to the top
of the fullscreen layer
2) When a view is focused, which isn't fullscreen, we need to make sure
all views from the fullscreen layer come back to the workspace layer, and
then put this view on top of them.
3) When a view is removed from a layer, we need to check if the topmost
view hasn't become a fullscreen surface. If so, just raise it to the
fullscreen layer.

Doing those things ensures that:
1) there are views in the fs layer only if there isn't a regular view on top
2) any topmost fs surface is in the fullscreen layer