cube: Add skybox and rotate keybindings
This commit is essentially a cube rewrite that adds skydome
and cubemap implementations to use an image as cube background.
Rotation is a bit different, now there is a zoom option that
determines how far away from the camera cube will zoom when
free rotating. Also now available are keybindings to rotate
cube left and right.