cube: Fix cube when no skydome image is set
When skydome was selected with no image set, the cube would not
activate at all. This was partially because the background methods
were clearing the depth buffer and skydome was only doing this
when there was a texture set. This makes it so the background is
cleared to green when skydome or cubemap are selected but the
texture for them are not loaded.