animate: ensure only a single animation per view is ever running
Otherwise we get into race conditions: if a view maps, unmaps and maps
again (doing so in a very short amount of time), we'd have got both
unmap and map animations at the same time. This could lead to a crash,
as both of those animations could use the same fire transformer.