cursor: send button events back to wlroots even with an active grab
We need this so that wlroots doesn't think we have a pressed button
while there isn't (needed for DnD). Nevertheless, the patch here might
still cause wlroots button counter to get corrupted, this time making it
go negative. We'd need wlroots to track pressed buttons, so that we can
avoid sending events in that case.