wm: add fini methods to wm plugins
We generally relied on the fact that an output would auto-clean all
bindings associated with it, but it is still better to clean up things.
This also fixes a crash in set_focus, where we connect a signal to a
view and don't reset it when the plugin is destroyed.