Fix race condition when forking child to background
Sometimes, calling waitpid() would wait forever in the case where exit(0)
was called first. This was the case on amd cpu when switching to another
workspace and attempting to launch an app with keybinding. We actually
don't need this call to avoid zombie processes since we are already using
the double fork method. Fixes #224.