core: rework tile, maximize and fullscreen APIs
With this change, core is responsible for keeping the last windowed and
last maximized view geometry, so that the grid plugin isn't necessary to
have maximize/fullscreen functionality.

When the maximize and fullscreen requests are sent, the desired size is
calculated by core, and grid uses it instead of manually storing this
data.

Lastly, the maximized state has been removed completely since it
overlaps with tiled state. To maximize a window, simply tile it to all
the 4 edges of the screen, to unmaximize, tile to no edges.