move-snap-helper: Ensure view stays around the grab if resized
The view size may change while it is being moved (i.e when it is
unsnapped), however we want to keep the view centered.

In this way, we can also remove the special casing in grid where it
tries to detect whether move is active, because it can now restore the
view independetly from move, and move will move the view to the correct
place.