xwayland: Use configure request mask to position surfaces
The mask contains information about whether or not to use the
position and size fields. Since size is determined by commit,
we only concern ourselves with the position here. This requires
wlroots #1843. This fixes bugs with positioning xwayland surfaces
with and without place plugin enabled. Fixes #307. Thanks to
ammen99 for the initial work on this.