view: fix wlr_surface usage due to upstream change
There was a change in the definition of the wlr_surface struct, which
led to breakage in the surface.cpp code. This commit fixes the
issue in wlr_surface_base_t::get_buffer(), which was the only affected
place (at least in regards to compile-time errors).

This change was introduced in wlroots commit 8afc1ed6, and briefly
discussed in pull request #2043.