render-manager: Add get_swap_damage function
There are at least two damage regions for a frame. We have client damage
which is the region of all surface damage for the output and the swap
damage which is the region passed to swap buffers. Typically we are
only interested in client damage but for purposes of plugins like
showrepaint, we need the swap damage to have a comprehensive view of
all damage happening on the output.