xwayland: fix configure_request handling for unmapped views
For unmapped views, we should not attempt to modify their configure
requests. This is also what Sway does.

Additionally, we should watch out for views which are not
override-redirect but do not have window type normal. These views should
be allowed to move around the screen as the client configures them.

Fixes #487