scale: Keep view detached signal connected until scale animation completes
If a view is closed while the unscale animation is running, wayfire could crash.
This is because we were disconnecting the view detached signal handler at the
beginning of scale animation and if a view is closed during this time, it was not
removed from the list, leaving a dufunct object. Fix this by waiting until
finalize to disconnect the detach signal.