workarounds: add support for non-modal dialogs.
Fixes #869

Wayland does not have the concept of modal and non-modal dialogs.
This might come as a future extension to the xdg-shell protocol.

In the meantime, we can provide users with an option to choose whether
dialogs should all be modal (as it was up to now) or non-modal.

If all dialogs are modal: whenever a dialog is open, it will always be
focused, and the main window will not be able to receive keyboard focus
until the dialog is closed.

If all dialogs are non-modal: dialogs are still kept on top of their
respective main window. However, the main window can still be focused.
Some toolkits like Qt apply certain hacks - Qt will ignore events on the
main window if it has opened a dialog it thinks is modal.

Overall, setting `workarounds/all_dialogs_modal` to `true` (which is the
        default) will keep the current behavior in Wayfire.

Setting it to `false` makes Wayfire behave like Kwin and Mutter,
practically forcing applications to take care of whether they have a
modal or a non-modal dialog.