scale: refactor layout_slots() logic
previously there was a lot of duplicated code because the main view and
children were handled separately. This is not really needed, so I
refactored it a bit and made animations for new child views in interactive
mode better.

Also added a few more comments so that the logic of the whole can be
followed easily.