scale: allow overlays to be added to the transformer
Such overlays can define a pre-render and render hook that will be called at the appropriate times in the rendering cycle (similarly to how pre-render and overlay effects can be added to the main renderer).

Overlays can define padding to be added around the view, e.g. to render outside it and also for scale to take into account when calculating the layout.