vswitch: add support for motion to absolute workspace pos (#1199)
There is also support for directly moving a view to a workspace.

Vswitch now supports the following options:

binding_{N, dir} -> Move to workspace N or in the given direction
with_win_{N, dir} -> Move to workspace N or in the given direction
together with focused window.
send_win_{N, dir} -> Move focused window to workspace N or in the given
direction.

N is the workspace number, workspaces are numbered from left to right,
  top to bottom.
dir can be left, right, up or down