workspace-manager: ignore promoted state for views not in WORKSPACE layer (#1333)
Promoted views make sense only for that layer. If a promoted view is
moved to some of the other layers, it may become inaccessible because
its promoted status is first cleared when it is moved back to the
workspace layer.

We can simply ignore the promoted state as long as the view is not in
the workspace layer. We also don't have to clear this state because it
is unused. It will be recalculated when the view is brought back to the
WORKSPACE layer.

Fixes #1076