api: split views and surfaces (#1370)
Instead of making view_interface_t a subclass of surface_interface_t, we
use composition. Now each view has a main surface accessible via
get_main_surface().

This allows for cleaner separation of responsibilities, and we don't
have to use hacks in wlr_surface_base_t to avoid double inheritance.

Also, this is a second step to supporting the same window being visible
on multiple outputs, as we use shared pointers for surfaces. Now, we can
have multiple views which use the same surfaces. This is of course also
useful for plugins which need to override the way Wayfire handles
surfaces - for example, by providing a custom Xwayland view
implementation, but reusing the wlr_surface handling. Another potential
example are client-rendered server side decorations, or implementations
of custom shells.

Preparation for the final part of #1309, which is splitting the desktop
surface part out of the view.