Implement switch to last workspace feature for vswitch plugin (#1371)
Extend the vswitch plugin with bindings that allow switching back to the
last active workspace. The functionality is achieved by always
remembering the switching direction at each switch, so that we can
easily reverse it. There is not only a binding for switching to the *last*
workspace, but also two other bindings: 1) for taking the current view
with when switching and 2) for only sending the currently focused view
to the *last* workspace.