Add an option "rules" to window-rules with type "dynamic-list" (#1607)
Currently, window-rules does not have any option. Therefore, it's difficult for the
parser to auto-magically load the stored rules. Furthermore, since it does not have
any option defined, wcm also does not list the rules, or allow it's modification.
This MR adds the "rules" option of type "dynamic-list" to window-rules.xml file.