scale: Fix title overlay
Since we have a grab over the scale views, attempting to get the
view under the cursor with plugin API functions fails. So, we use
our own function tailored for this purpose. The scale implementation
already uses this function. Move it to a header inlined to share
in with the other scale source files. Fixes #1631.