api: refactor view initialization sequence and use shared_ptr ownwership
Up to now, views were using a custom made scheme based on take_ref and
unref, which is manual reference counting. In addition, we were using
core for initialization of views and keeping track of which views are
allocated.

Instead, we can use the new tracking_allocator to keep track of all
views which have been allocated, and we do not need to manually add and
remove views from core.

In addition, the usage of shared_ptr also allows the use of weak_ptr
for simple handling of views disappearing.

Lastly, the views have been refactored to use a static factory method
for allocation and initialization. This replaces the old system with
the virtual initialize() method, which was less obvious to use.

Fixes #1369